requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
google-generativeai==0.5.0
pydantic==2.8.0
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0